// projects/basket/app/client/pages/BasketPage/analytics/helpers/getProductAnalyticData/getProductAnalyticsData.types.ts
export interface ProductAnalyticsData {
  productPackageId: string[]
  productBrand: string[]
  productCapacity: string[]
  productColour: string[]
  productId: string[]
  productModel: string[]
  productName: string[]
  productPlanData: string[]
  productPackagePlanDuration: string[]
  productPlanMinutes: string[]
  productPlanSms: string[]
  productPrice: string[]
  productPriceNet: string[]
  productPriceRecurrenceAmount: string[]
  productPriceRecurrenceAmountNet: string[]
  productCategory: string[]
  productPackageNetworkSwitchType: string[]
  productDiscountValue: string[]
  productDiscountValueNet: string[]
  productDiscountRecurrenceAmount: string[]
  productDiscountRecurrenceAmountNet: string[]
  productPackageName: string[]
  productQuantity: string[]
  transactionCouponCode: BasketV2.Basket['voucherCode']
  quoteId: string[]
  tradeInDuration: number[]
  tradeInType: string[]
  tradeInAmount: number[]
  tradeInDevice: string[]
}